* Fix some warnings * Codeblocks projects debug targets now compile with -Wall * Make destructor virtual in ServerBase class * Update targets in cb projects to be set up with platform specific targets (still needs work) * Comment out some unused code
SVN r153 (trunk)

+++ server/src/sv_master.cpp
@@ -394,7 +394,7 @@ void SV_SendServerInfo()
 //bond===========================
 	MSG_WriteString(&ml_message, (char *)email.cstring());
 
-	int timeleft = timelimit - level.time/(TICRATE*60);
+	int timeleft = (int)(timelimit - level.time/(TICRATE*60));
 	if (timeleft<0) timeleft=0;
 
 	MSG_WriteShort(&ml_message,(int)timelimit);
